+++ RUFAS/biophysical/animal/herd_manager.py
@@ -25,6 +25,7 @@ from RUFAS.biophysical.animal.pen import Pen
 from RUFAS.biophysical.animal.ration.calf_ration_manager import CalfMilkType, CalfRationManager, WHOLE_MILK_ID
 from RUFAS.biophysical.animal.ration.user_defined_ration_manager import UserDefinedRationManager
 from RUFAS.current_day_conditions import CurrentDayConditions
+from RUFAS.data_structures.animal_manure_excretions import AnimalManureExcretions
 from RUFAS.data_structures.animal_to_manure_connection import ManureStream
 from RUFAS.data_structures.feed_storage_to_animal_connection import (
     Feed,
@@ -505,17 +506,18 @@ class HerdManager:
 
         self.record_pen_history(time.simulation_day)
 
+        animal_manure_excretions_by_pen: dict[str, AnimalManureExcretions] = {}
         herd_manager_output: dict[str, ManureStream] = {}
+        enteric_methane_emission_by_pen: dict[str, float] = {}
         for pen in self.all_pens:
+            animal_manure_excretions_by_pen[f"{pen.animal_combination.name}_PEN_{pen.id}"] = pen.total_manure_excretion
             herd_manager_output.update(pen.get_manure_streams())
-
-        enteric_methane_emission_by_pen: dict[str, float] = {
-            f"{pen.id}_{pen.animal_combination.name}": pen.total_enteric_methane for pen in self.all_pens
-        }
+            enteric_methane_emission_by_pen[f"{pen.animal_combination.name}_PEN_{pen.id}"] = pen.total_enteric_methane
 
         self.update_herd_statistics()
 
-        # AnimalModuleReporter.report_animal_module_manure(herd_manager_output)
+        AnimalModuleReporter.report_manure_excretions(animal_manure_excretions_by_pen, time.simulation_day)
+        AnimalModuleReporter.report_manure_streams(herd_manager_output, time.simulation_day)
         AnimalModuleReporter.report_enteric_methane_emission(enteric_methane_emission_by_pen)
         AnimalModuleReporter.report_daily_reports(self, time.simulation_day)
 
@@ -818,18 +820,7 @@ class HerdManager:
             minutes_away_for_milking = pen_data.get("minutes_away_for_milking", 120)
             first_parlor_processor = pen_data.get("first_parlor_processor", None)
             parlor_stream_name = pen_data.get("parlor_stream_name", None)
-            manure_streams = pen_data.get(
-                "manure_streams",
-                # TODO remove this default value when metadata properties are updated in issue #2272
-                [
-                    {
-                        "stream_name": "general_pen",
-                        "bedding_name": "sand",
-                        "stream_proportion": 1.0,
-                        "first_processor": "general_handler",
-                    }
-                ],
-            )
+            manure_streams = pen_data.get("manure_streams")
 
             pen = Pen(
                 pen_id=pen_id,